requirements.txt
dash
flask
plotly
python-dotenv
requests
xmltodict
